feat: Add user module and controller (#2)
* feat: Add user module and controller

The code changes include adding the user module and controller, along with the necessary DTOs for creating and updating users. This commit introduces the basic structure for managing users in the application.

Note: The recent repository commits suggest a convention of using a prefix to indicate the type of changes made in the commit message.

* feat: Add client_api_key column to users table

This commit adds a new column "client_api_key" to the "users" table in the database. The column is of type varchar and has a default value of gen_random_uuid(). It is marked as NOT NULL. Additionally, a unique constraint "users_client_api_key_unique" is added to ensure the uniqueness of the client API keys.

Note: The recent repository commits suggest a convention of using a prefix to indicate the type of changes made in the commit message.

---------

Co-authored-by: Akash-Manikandan <akash2003m@gmail.com>